Rename FastqCorrector.sh to FastqCorrector_0.9.sh
Uses a provided file with read IDs (format as Tablet assembly visualizer copies read information to clipboard) to search the coordinates of each read in a Velvet "Sequences" file (produced as the result of an assembly of raw reads). Afterwards it searches those read coordinates in raw FASTQ files to delete them and produce new FASTQ files ready for Velvet assembly.