Rename genome_feature_stats.sh to genomes_feature_stats.sh
From a given set of feature tables it calculates the mean average feature lengths and prints all the feature of length below the threshold (the average calculated) to an output file.  Features consisting of CDS are also printed with its corresponding protein product name. Locus tag is printed for each feature to make further analyses easier.